v1.2.0: new methods override: on message/edited, fix convo
use a custom dispatcher for on_message and on_edited_message
convo: check if self is in the list before attempting to remove
load: add docstring and response before downloading plugin.

+++ ub_core/core/handlers/handler.py
@@ -1,7 +1,7 @@
 import asyncio
 from typing import Callable
 
-from pyrogram import StopPropagation
+from pyrogram import ContinuePropagation, StopPropagation
 from pyrogram.handlers import EditedMessageHandler, MessageHandler
 from pyrogram.types import Message as Msg
 
@@ -14,6 +14,8 @@ async def cmd_dispatcher(
     message: Message,
     func: Callable = None,
     check_for_reactions: bool = True,
+    mode_sensitive: bool = True,
+    is_command: bool = False,
 ) -> None:
     """Custom Command Dispatcher to Gracefully Handle Errors and Cancellation"""
     if check_for_reactions and filters.anti_reaction(message):
@@ -32,15 +34,20 @@ async def cmd_dispatcher(
 
     try:
         await task
-        if message.is_from_owner:
+        if is_command and message.is_from_owner:
             await message.delete()
+
     except asyncio.exceptions.CancelledError:
         await client.log_text(text=f"<b>#Cancelled</b>:\n<code>{message.text}</code>")
-    except StopPropagation:
+
+    except (StopPropagation, ContinuePropagation):
         raise
+
     except Exception as e:
         client.log.error(e, exc_info=True, extra={"tg_message": message})
-    message.stop_propagation()
+
+    if is_command:
+        message.stop_propagation()
 
 
 # Don't Load Handler is Value is not True
@@ -57,8 +64,8 @@ if Config.LOAD_HANDLERS:
     )
 
 
-@bot.on_message(filters.convo_filter, group=0)
-@bot.on_edited_message(filters.convo_filter, group=0)
+@bot.on_message(filters.convo_filter, group=0, is_command=False)
+@bot.on_edited_message(filters.convo_filter, group=0, is_command=False)
 async def convo_handler(bot: BOT, message: Msg):
     """Check for convo filter and update convo future accordingly"""
     conv_objects: list[Convo] = Convo.CONVO_DICT[message.chat.id]
